ml-service/requirements.txt
# MediScan ML Service — Dependencies
# ====================================

# Web framework
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6

# ML / Deep Learning
torch>=2.0.0
torchvision>=0.15.0

# OCR
easyocr>=1.7.0
pytesseract>=0.3.10

# Image processing
opencv-python-headless>=4.8.0
Pillow>=10.0.0
numpy>=1.24.0

# Text matching
rapidfuzz>=3.5.0

# Text detection
craft-text-detector>=0.4.3

# Database
mysql-connector-python>=8.2.0

# Data processing
pandas>=2.0.0
scikit-learn>=1.3.0

# Visualization (training/evaluation only)
matplotlib>=3.7.0
seaborn>=0.12.0
